Move SecretStoreResource and Resource to their own files
