+++ crates/slightfile/src/lib.rs
@@ -3,6 +3,10 @@ use std::{collections::HashMap, fmt::Display, path::Path};
 
 use serde::{Deserialize, Deserializer, Serialize};
 
+pub mod resource;
+pub mod secret_store;
+pub use resource::Resource;
+pub use secret_store::SecretStoreResource;
 /// slightfile version.
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
 pub enum SpecVersion {
@@ -107,134 +111,6 @@ pub struct Config {
 impl Config {
     pub fn new(name: String, value: String) -> Self {
         Self { name, value }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum SecretStoreResource {
-    #[serde(rename = "configs.azapp")]
-    Azapp,
-    #[serde(rename = "configs.envvars")]
-    Envvars,
-    #[serde(rename = "configs.usersecrets")]
-    Usersecrets,
-    #[serde(rename = "configs.local")]
-    Local,
-}
-
-impl TryFrom<String> for SecretStoreResource {
-    type Error = anyhow::Error;
-
-    fn try_from(value: String) -> Result<Self, Self::Error> {
-        match value.as_str() {
-            "configs.azapp" => Ok(SecretStoreResource::Azapp),
-            "configs.envvars" => Ok(SecretStoreResource::Envvars),
-            "configs.usersecrets" => Ok(SecretStoreResource::Usersecrets),
-            "configs.local" => Ok(SecretStoreResource::Local),
-            _ => bail!("Unknown secret store resource: {}", value),
-        }
-    }
-}
-
-impl From<SecretStoreResource> for String {
-    fn from(value: SecretStoreResource) -> Self {
-        match value {
-            SecretStoreResource::Azapp => String::from("configs.azapp"),
-            SecretStoreResource::Envvars => String::from("configs.envvars"),
-            SecretStoreResource::Usersecrets => String::from("configs.usersecrets"),
-            SecretStoreResource::Local => String::from("configs.local"),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum Resource {
-    #[serde(rename = "blobstore.aws_s3")]
-    BlobstoreAwsS3,
-    #[serde(rename = "blobstore.azblob")]
-    BlobstoreAzblob,
-    #[serde(rename = "keyvalue.awsdynamodb")]
-    KeyvalueAwsDynamoDb,
-    #[serde(rename = "keyvalue.azblob")]
-    KeyvalueAzblob,
-    #[serde(rename = "keyvalue.filesystem")]
-    KeyvalueFilesystem,
-    #[serde(rename = "keyvalue.redis")]
-    KeyvalueRedis,
-    #[serde(rename = "kv.awsdynamodb")]
-    V1KeyvalueAwsDynamoDb,
-    #[serde(rename = "kv.azblob")]
-    V1KeyvalueAzblob,
-    #[serde(rename = "kv.filesystem")]
-    V1KeyvalueFilesystem,
-    #[serde(rename = "kv.redis")]
-    V1KeyvalueRedis,
-    #[serde(rename = "keyvalue.dapr")]
-    KeyvalueDapr,
-    #[serde(rename = "messaging.azsbus")]
-    MessagingAzsbus,
-    #[serde(rename = "messaging.confluent_apache_kafka")]
-    MessagingConfluentApacheKafka,
-    #[serde(rename = "messaging.filesystem")]
-    MessagingFilesystem,
-    #[serde(rename = "messaging.mosquitto")]
-    MessagingMosquitto,
-    #[serde(rename = "mq.azsbus")]
-    V1MessagingAzsbus,
-    #[serde(rename = "mq.filesystem")]
-    V1MessagingFilesystem,
-    #[serde(rename = "messaging.nats")]
-    MessagingNats,
-    #[serde(rename = "http")] // TODO: change this to http-server and bump up slightfile version?
-    HttpServer,
-    #[serde(rename = "http-client")]
-    HttpClient,
-    #[serde(rename = "configs.azapp")]
-    ConfigsAzapp,
-    #[serde(rename = "configs.envvars")]
-    ConfigsEnvvars,
-    #[serde(rename = "configs.usersecrets")]
-    ConfigsUsersecrets,
-    #[serde(rename = "distributed_locking.etcd")]
-    DistributedLockingEtcd,
-    #[serde(rename = "lockd.etcd")]
-    V1DistributedLockingEtcd,
-    #[serde(rename = "sql.postgres")]
-    SqlPostgres,
-}
-
-impl Display for Resource {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Resource::BlobstoreAwsS3 => write!(f, "blobstore.aws_s3"),
-            Resource::BlobstoreAzblob => write!(f, "blobstore.azblob"),
-            Resource::KeyvalueAwsDynamoDb => write!(f, "keyvalue.awsdynamodb"),
-            Resource::KeyvalueAzblob => write!(f, "keyvalue.azblob"),
-            Resource::KeyvalueFilesystem => write!(f, "keyvalue.filesystem"),
-            Resource::KeyvalueRedis => write!(f, "keyvalue.redis"),
-            Resource::KeyvalueDapr => write!(f, "keyvalue.dapr"),
-            Resource::MessagingAzsbus => write!(f, "messaging.azsbus"),
-            Resource::MessagingConfluentApacheKafka => {
-                write!(f, "messaging.confluent_apache_kafka")
-            }
-            Resource::MessagingFilesystem => write!(f, "messaging.filesystem"),
-            Resource::MessagingMosquitto => write!(f, "messaging.mosquitto"),
-            Resource::MessagingNats => write!(f, "messaging.nats"),
-            Resource::HttpServer => write!(f, "http"),
-            Resource::HttpClient => write!(f, "http-client"),
-            Resource::ConfigsAzapp => write!(f, "configs.azapp"),
-            Resource::ConfigsEnvvars => write!(f, "configs.envvars"),
-            Resource::ConfigsUsersecrets => write!(f, "configs.usersecrets"),
-            Resource::DistributedLockingEtcd => write!(f, "distributed_locking.etcd"),
-            Resource::SqlPostgres => write!(f, "sql.postgres"),
-            Resource::V1KeyvalueAwsDynamoDb => write!(f, "kv.awsdynamodb"),
-            Resource::V1KeyvalueAzblob => write!(f, "kv.azblob"),
-            Resource::V1KeyvalueFilesystem => write!(f, "kv.filesystem"),
-            Resource::V1KeyvalueRedis => write!(f, "kv.redis"),
-            Resource::V1MessagingAzsbus => write!(f, "mq.azsbus"),
-            Resource::V1MessagingFilesystem => write!(f, "mq.filesystem"),
-            Resource::V1DistributedLockingEtcd => write!(f, "lockd.etcd"),
-        }
     }
 }
